+++ 10-binary_tree_depth.c
@@ -1,0 +1,25 @@
+#include "binary_trees.h"
+#include <stdio.h>
+#include <stdio.h>
+
+/**
+ * binary_tree_depth - Measures the depth of a node in a binary tree.
+ * @tree: Pointer to the node to measure the depth.
+ * Return: Depth of the node. If tree is NULL, return 0.
+ */
+size_t binary_tree_depth(const binary_tree_t *tree)
+{
+	size_t depth = 0;
+
+	if (tree == NULL)
+	{
+		return (0);
+	}
+	while (tree->parent != NULL)
+	{
+		depth++;
+		tree = tree->parent;
+	}
+
+	return (depth);
+}
