+++ binary_trees.h
@@ -25,4 +25,5 @@ binary_tree_t *binary_tree_insert_right(binary_tree_t *parent, int value);
 int binary_tree_is_leaf(const binary_tree_t *node);
 void binary_tree_preorder(const binary_tree_t *tree, void (*func)(int));
 void binary_tree_postorder(const binary_tree_t *tree, void (*func)(int));
+size_t binary_tree_depth(const binary_tree_t *tree);
 #endif
